refactor(todo.js, dragManager.js): 드래그 관련 로직을 분리
