+++ scripts/components/todo.js
@@ -1,37 +1,27 @@
 import { TodoItem } from '../domain/todoItem.js'
-import { classNames, templateNames } from '../strings.js'
+import { classNames, keys, templateNames } from '../strings.js'
 import {
     createDomElementAsChild,
-    createDomElementAsSibling,
     findDomElement,
     removeDomElement,
     replaceDomElement,
 } from '../utils/domUtil.js'
 import { getState, setState } from '../utils/stateUtil.js'
 import { loadData, storeData } from '../utils/storageUtil.js'
-import { Category } from '../domain/category.js'
-
-// const TODO_LIST_STORAGE_KEY = 'todoList'
-const TODO_CATEGORY_KEY = 'todoCategory'
-// const TODO_FORM_DOM_ID_KEY = 'isCreatingTodo'
-const DRAG_ELEMENT_KEY = 'dragElement'
+import { manageDragEvents, manageDropEvents } from './dragManager.js'
 
 export const initTodo = () => {
-    let categoryList = loadData(TODO_CATEGORY_KEY)
+    let categoryList = loadData(keys.TODO_CATEGORY_KEY)
     if (!categoryList) {
-        storeData(TODO_CATEGORY_KEY, [])
+        storeData(keys.TODO_CATEGORY_KEY, [])
         categoryList = []
     }
 
     // 더미 데이터
     // categoryList = [Category(-1, '해야할 일'), Category(-1, '하고 있는 일')]
-    // storeData(TODO_CATEGORY_KEY, categoryList)
+    // storeData(keys.TODO_CATEGORY_KEY, categoryList)
 
     const mainElement = document.querySelector('.main')
-    let prevCategory = null
-    let prevIndex = null
-    let currentCategory = null
-    let currentIndex = null
 
     categoryList.map((category) => {
         const categoryId = createDomElementAsChild(
@@ -55,154 +45,7 @@ export const initTodo = () => {
                         console.log(category.values.todoList)
                     })
 
-                // ###### 드래그 이벤트 ######
-                component.addEventListener('dragover', (e) => {
-                    // 드롭을 허용하기 위해 기본 동작 취소
-                    e.preventDefault()
-                })
-                component.addEventListener('drop', (e) => {
-                    // 일부 요소의 링크 열기와 같은 기본 동작 취소
-                    e.preventDefault()
-                    setState(DRAG_ELEMENT_KEY, null)
-                })
-
-                let dragDepth = 0
-                component.addEventListener('dragenter', (e) => {
-                    dragDepth++
-                    const [originElementId, originIndex] =
-                        getState(DRAG_ELEMENT_KEY)
-
-                    const ghostElement = findDomElement(originElementId)
-                    let updateFlag = false
-                    const categoryList = getState(TODO_CATEGORY_KEY)
-                    for (let category of categoryList) {
-                        // 카테고리 식별
-                        const parentElement = findDomElement(
-                            category.identifier
-                        )
-                        if (!parentElement.contains(e.target)) continue
-                        prevCategory = currentCategory ?? category
-                        currentCategory = category
-                        currentCategory.identifier
-                        // )
-                        if (
-                            prevCategory.identifier !==
-                            currentCategory.identifier
-                        ) {
-                            // 카테고리가 바뀌면 skeleton update
-                            updateFlag = true
-                        }
-
-                        // todoList의 몇 번째 index인지 식별
-                        for (let [
-                            idx,
-                            todoItem,
-                        ] of category.values.todoList.entries()) {
-                            if (todoItem.identifier === e.target.id) {
-                                prevIndex = currentIndex ?? originIndex
-                                currentIndex = idx
-                                if (prevIndex !== currentIndex) {
-                                    // index가 바뀌면 skeleton update
-                                    updateFlag = true
-                                }
-                                break
-                            }
-                        }
-
-                        let isInDragZone = false
-                        if (
-                            e.target.classList.contains(classNames.todoDragZone)
-                        ) {
-                            if (currentCategory.values.todoList.length > 0) {
-                                currentIndex =
-                                    currentCategory.values.todoList.length - 1
-                            } else {
-                                currentIndex = 0
-                            }
-                            updateFlag = true
-                            isInDragZone = true
-                        }
-
-                        if (!updateFlag) return
-
-                        const targetElement =
-                            currentCategory.values.todoList.length > 0
-                                ? findDomElement(
-                                      currentCategory.values.todoList[
-                                          currentIndex
-                                      ].identifier
-                                  )
-                                : null
-
-                        if (prevCategory !== currentCategory) {
-                            if (isInDragZone) {
-                                if (targetElement === null) {
-                                    component
-                                        .querySelector(
-                                            `.${classNames.todoBody}`
-                                        )
-                                        .appendChild(ghostElement)
-                                } else {
-                                    targetElement.after(ghostElement)
-                                }
-                                currentCategory.values.todoList.push(
-                                    prevCategory.values.todoList[prevIndex]
-                                )
-                                prevCategory.values.todoList.splice(
-                                    prevIndex,
-                                    1
-                                )
-                                return
-                            }
-                            targetElement.before(ghostElement)
-                            const originTodoItem =
-                                prevCategory.values.todoList[prevIndex]
-                            const todoList = currentCategory.values.todoList
-                            prevCategory.values.todoList.splice(prevIndex, 1)
-                            currentCategory.values.todoList = [
-                                ...todoList.slice(0, currentIndex),
-                                originTodoItem,
-                                ...todoList.slice(currentIndex),
-                            ]
-                            return
-                        }
-
-                        const todoList = category.values.todoList
-                        if (prevIndex < currentIndex) {
-                            targetElement.after(ghostElement)
-                            category.values.todoList = [
-                                ...todoList.slice(0, prevIndex),
-                                ...todoList.slice(
-                                    prevIndex + 1,
-                                    currentIndex + 1
-                                ),
-                                todoList[prevIndex],
-                                ...todoList.slice(currentIndex + 1),
-                            ]
-                        } else if (prevIndex > currentIndex) {
-                            targetElement.before(ghostElement)
-                            category.values.todoList = [
-                                ...todoList.slice(0, currentIndex),
-                                todoList[prevIndex],
-                                ...todoList.slice(currentIndex, prevIndex),
-                                ...todoList.slice(prevIndex + 1),
-                            ]
-                        }
-                    }
-                })
-                component.addEventListener('dragleave', (e) => {
-                    dragDepth--
-                    if (dragDepth > 0) return
-                    console.log('dragleave', category.identifier)
-                })
-                component.addEventListener('drop', (e) => {
-                    console.log('drop', category.identifier)
-                    prevCategory = null
-                    prevIndex = null
-                    currentCategory = null
-                    currentIndex = null
-                    storeData(TODO_CATEGORY_KEY, getState(TODO_CATEGORY_KEY))
-                })
+                manageDropEvents(component, category)
             }
         )
         category.identifier = categoryId
@@ -227,7 +70,7 @@ export const initTodo = () => {
         return category
     })
 
-    setState(TODO_CATEGORY_KEY, categoryList)
+    setState(keys.TODO_CATEGORY_KEY, categoryList)
 }
 
 const onAddTodoButtonClick = (category) => {
@@ -304,26 +147,7 @@ const initTodoItemElement = (todoElement, todoItem) => {
         .addEventListener('click', () => {
             editTodoItem(identifier)
         })
-    manageDrag(todoElement.querySelector(`.${classNames.todoItemBody}`))
-}
-
-const manageDrag = (element) => {
-    // let isDragging = false
-    let currentCategory = null
-    let currentIndex = null
-    element.addEventListener('dragstart', (e) => {
-        // console.log(e.target)
-        // console.log(element.offsetHeight)
-        element.classList.add(classNames.afterImage)
-        const index = Array.from(element.parentNode.children).indexOf(element)
-        setState(DRAG_ELEMENT_KEY, [e.target.id, index])
-    })
-    element.addEventListener('drag', (e) => {
-        // console.log('drag', e.target)
-    })
-    element.addEventListener('dragend', (e) => {
-        element.classList.remove(classNames.afterImage)
-    })
+    manageDragEvents(todoElement.querySelector(`.${classNames.todoItemBody}`))
 }
 
 const addTodoItem = (title, content, author, category) => {
@@ -342,12 +166,12 @@ const addTodoItem = (title, content, author, category) => {
     )
 
     // category 객체를 참조하므로 setState를 안 해도 변경이 되긴 함 .. 맘에 안들지만 일단은 이렇게
-    storeData(TODO_CATEGORY_KEY, getState(TODO_CATEGORY_KEY))
+    storeData(keys.TODO_CATEGORY_KEY, getState(keys.TODO_CATEGORY_KEY))
     renewTodoCount(category)
 }
 
 const getTodoItemInfo = (identifier) => {
-    const categoryList = getState(TODO_CATEGORY_KEY)
+    const categoryList = getState(keys.TODO_CATEGORY_KEY)
     for (let category of categoryList) {
         for (let [idx, todo] of category.values.todoList.entries()) {
             if (todo.identifier === identifier) {
@@ -363,7 +187,7 @@ const removeTodoItem = (identifier) => {
     category.values.todoList.splice(index, 1)
     removeDomElement(identifier)
     renewTodoCount(category)
-    storeData(TODO_CATEGORY_KEY, getState(TODO_CATEGORY_KEY))
+    storeData(keys.TODO_CATEGORY_KEY, getState(keys.TODO_CATEGORY_KEY))
 }
 
 const editTodoItem = (identifier) => {
@@ -415,7 +239,10 @@ const editTodoItem = (identifier) => {
                             todoList[targetIdx] = editedTodoItem
                         }
                     )
-                    storeData(TODO_CATEGORY_KEY, getState(TODO_CATEGORY_KEY))
+                    storeData(
+                        keys.TODO_CATEGORY_KEY,
+                        getState(keys.TODO_CATEGORY_KEY)
+                    )
                 })
             component
                 .querySelector(`.${classNames.todoEditFormCancelBtn}`)
